bootcamp ownership security update, delete, upload photo

+++ controllers/bootcamps.js
@@ -71,10 +71,7 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
  * @access Private
  */
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-	const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-		new: true,
-		runValidators: true,
-	});
+	let bootcamp = await Bootcamp.findById(req.params.id);
 	if (!bootcamp)
 		return next(
 			new ErrorResponse(
@@ -83,6 +80,18 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 			),
 		);
 
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`User ${req.params.id} is not authroized to update this bootcamp`,
+        401)) 
+    }
+    bootcamp = await Bootcamp.findOneAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true
+    });
 	res.status(200).json({ success: true, data: bootcamp });
 });
 
@@ -99,6 +108,15 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 				404,
 			),
 		);
+
+    // make sure user is bootcamp owner
+    if(bootcamp.user.tostring() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`user ${req.params.id} is not authroized to delete this bootcamp`,
+                401)) 
+    }
 
 	bootcamp.deleteOne();
 	res.status(200).json({
@@ -152,6 +170,15 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 			),
 		);
 	}
+    // Make sure user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id &&
+        req.user.role !== 'admin'
+    ) {
+        return next( 
+            new ErrorResponse(`User ${req.params.id} is not authroized to update this bootcamp`,
+                401)) 
+    }
+
 	if (!req.files) {
 		return next(new ErrorResponse(`Please upload a file`, 400));
 	}
